chore(grz-common,grz-cli,grzctl,grz-db,grz-pydantic-models): Upgrade pydantic to 2.12 (#425)
Upgrades pydantic (and subsequently pydantic-settings, sqlmodel an
sqlalchemy) to 2.12.
This necessitates some changes in type annotations, as there were some
behaviour changes regarding inheritance and generics, as well as some
projected deprecations (model_serializer in the model config replaced by
custom field serializers).
Also relaxes the upper bound of pydantic to be a little less strict,
hoping there won't be too many breaking changes between minor versions
in the future.

Note that the `__table_args__ = {"extend_existing": True}` seems to be
necessary for tests to run without a hitch; tried using
```py
@pytest.fixture(autouse=True)
def clear_sqlmodel_metadata():
    SQLModel.metadata.clear()
```
to work around that, but perhaps that's the wrong way to go about this
and we should have a closer look at
https://sqlmodel.tiangolo.com/tutorial/fastapi/tests/?h=clear#client-fixture
instead